working on the cleanup process

+++ projectRun.py
@@ -38,21 +38,24 @@ def cleanUp(dir):
         knownLangPath = os.path.join(dir, language)
         for learnLangDir in os.listdir(knownLangPath):
             print learnLangDir
-            learnLangPath = os.path.join(knownLangPath, learnLangDir)
-            for unit in os.listdir(learnLangPath):
-                unitDir = os.path.join(learnLangPath, unit)
-                if unit == "config.yaml":
-                    os.remove(unitDir)
-                else:
-                    for B4X in os.listdir(unitDir):
-                        print B4X
-                        b4xDir = os.path.join(unitDir, B4X)
-                        if B4X == "OUTPUT_B4Xs_1":
-                            os.rmdir(b4xDir)
-                        elif unit.endswith(".b4u"):
-                            os.remove(b4xDir)
-                        elif unit.endswith(".yaml"):
-                            os.remove(b4xDir)
+            if learnLangDir == "revision.txt":
+                pass
+            else:
+                learnLangPath = os.path.join(knownLangPath, learnLangDir)
+                for unit in os.listdir(learnLangPath):
+                    unitDir = os.path.join(learnLangPath, unit)
+                    if unit == "config.yaml":
+                        os.remove(unitDir)
+                    else:
+                        for B4X in os.listdir(unitDir):
+                            print B4X
+                            b4xDir = os.path.join(unitDir, B4X)
+                            if B4X == "OUTPUT_B4Xs_1":
+                                os.rmdir(b4xDir)
+                            elif unit.endswith(".b4u"):
+                                os.remove(b4xDir)
+                            elif unit.endswith(".yaml"):
+                                os.remove(b4xDir)
 
 #dir = project level directory
 def checkNumberOfUnits(dir, projUnitNameList):
